requirements.txt
django_dropbox_storage==0.4.1
django-cleanup==2.1.0
Pillow==5.2.0
whitenoise==3.3.1
